extracted dcc page creation to separate service
